chat_setup/simple_adk_agent/requirements.txt
google-adk
python-dotenv
fastapi
uvicorn